+++ src/app/handlers/vcState_update_handler.ts
@@ -4,8 +4,8 @@ import { env } from '@/config/env';
 import { endEventOnRecruiterLeave } from '@/features/recruit/vc_reservation/event_auto_end';
 import { endCall, startCall } from '@/features/stats/voice_count';
 import { autokill } from '@/features/utils/voice/tts/discordjs_voice';
-import { vcToolsStickyFromVoiceState } from '@/features/vc_tools/vc_tools_message';
-import { disableLimit } from '@/features/vc_tools/voice_lock';
+import { vcToolsStickyFromVoiceState } from '@/features/utils/voice/vc_tools/vc_tools_message';
+import { disableLimit } from '@/features/utils/voice/vc_tools/voice_lock';
 import { log4js_obj } from '@/infra/logging/log4js';
 import { sendErrorLogs } from '@/infra/logging/send_error_logs';
 import { exists, notExists } from '@/shared/assert';
