--- src/features/vc_tools/index.ts
@@ -1,5 +1,0 @@
-/**
- * vc_tools の公開API。
- * 他の feature からはこのファイル経由でのみ参照する。
- */
-export { sendVCToolsSticky } from './vc_tools_message';
